refactor: extract authentication logic into dedicated useAuth hook
